+++ sdk/src/lib.rs
@@ -13,6 +13,7 @@ pub mod arithmetic;
 pub mod builtins;
 pub mod client;
 pub mod commitment_config;
+pub mod derivation_path;
 pub mod deserialize_utils;
 pub mod entrypoint;
 pub mod entrypoint_deprecated;
